Task3_ArxivChatbot/.ipynb_checkpoints/requirements-checkpoint.txt
streamlit
pandas
sentence-transformers
faiss-cpu
transformers
torch
scikit-learn
nltk
